Split app and cdn stacks
