+++ cdk/lib/app-stack.ts
@@ -6,13 +6,7 @@ import * as ecsPatterns from "aws-cdk-lib/aws-ecs-patterns";
 import * as ecrAssets from "aws-cdk-lib/aws-ecr-assets";
 import * as logs from "aws-cdk-lib/aws-logs";
 import * as ssm from "aws-cdk-lib/aws-ssm";
-import * as cloudfront from "aws-cdk-lib/aws-cloudfront";
-import * as origins from "aws-cdk-lib/aws-cloudfront-origins";
-import * as route53 from "aws-cdk-lib/aws-route53";
-import * as targets from "aws-cdk-lib/aws-route53-targets";
-import * as acm from "aws-cdk-lib/aws-certificatemanager";
 import * as iam from "aws-cdk-lib/aws-iam";
-import { HttpsRedirect } from "aws-cdk-lib/aws-route53-patterns";
 
 interface Props extends cdk.StageProps {
   db: cdk.aws_rds.DatabaseInstance;
@@ -21,9 +15,9 @@ interface Props extends cdk.StageProps {
 }
 
 export class AppStack extends cdk.Stack {
-  public readonly appLoadBalancerDNS: cdk.CfnOutput;
   public readonly appPort: number = 3000;
   public readonly cloudMapNamespace = "service.internal";
+  public readonly loadbalancer: cdk.aws_elasticloadbalancingv2.ApplicationLoadBalancer;
 
   constructor(scope: Construct, id: string, props: Props) {
     super(scope, id, props);
@@ -35,38 +29,10 @@ export class AppStack extends cdk.Stack {
       1
     );
 
-    const hostedZoneId = ssm.StringParameter.valueForStringParameter(
-      this,
-      `/env/hostedZoneId`,
-      1
-    );
-
     const cluster = new ecs.Cluster(this, "ServiceCluster");
 
     cluster.addDefaultCloudMapNamespace({ name: this.cloudMapNamespace });
 
-    const appService = this.createFargateService(
-      cluster,
-      domainName,
-      hostedZoneId,
-      db,
-      bucket,
-      !!production
-    );
-
-    this.appLoadBalancerDNS = new cdk.CfnOutput(this, "AppLoadBalancerDNS", {
-      value: appService.loadBalancer.loadBalancerDnsName,
-    });
-  }
-
-  private createFargateService(
-    cluster: ecs.Cluster,
-    domainName: string,
-    hostedZoneId: string,
-    db: cdk.aws_rds.DatabaseInstance,
-    bucket: cdk.aws_s3.Bucket,
-    production: boolean
-  ) {
     if (!db.secret) throw new Error("DB Secret required");
     const wwwDomainName = `www.${domainName}`;
 
@@ -194,72 +160,6 @@ export class AppStack extends cdk.Stack {
       "app-inbound"
     );
 
-    const zone = route53.HostedZone.fromHostedZoneAttributes(this, "MyZone", {
-      hostedZoneId,
-      zoneName: domainName,
-    });
-
-    const certificate = new acm.DnsValidatedCertificate(this, "Certificate", {
-      domainName,
-      subjectAlternativeNames: [`*.${domainName}`],
-      hostedZone: zone,
-      region: "us-east-1",
-    });
-
-    const cf = new cloudfront.Distribution(this, "myDist", {
-      domainNames: [wwwDomainName],
-      certificate,
-      defaultBehavior: {
-        viewerProtocolPolicy: cloudfront.ViewerProtocolPolicy.REDIRECT_TO_HTTPS,
-        allowedMethods: cloudfront.AllowedMethods.ALLOW_ALL,
-        origin: new origins.LoadBalancerV2Origin(fargateService.loadBalancer, {
-          protocolPolicy: cloudfront.OriginProtocolPolicy.HTTP_ONLY,
-        }),
-        functionAssociations: [
-          {
-            function: new cloudfront.Function(this, "Function", {
-              code: cloudfront.FunctionCode.fromInline(`
-                function handler(event) {
-                  var host = event.request.headers.host.value;
-                  var request = event.request;
-                  if (host.includes("cloudfront")) {
-                    var response = {
-                      statusCode: 404,
-                      statusDescription: "Not found",
-                    };
-                    return response;
-                  }
-                  return request;
-                }                  
-              `),
-            }),
-            eventType: cloudfront.FunctionEventType.VIEWER_REQUEST,
-          },
-        ],
-      },
-    });
-
-    new route53.ARecord(this, "SiteAliasRecord", {
-      recordName: wwwDomainName,
-      target: route53.RecordTarget.fromAlias(new targets.CloudFrontTarget(cf)),
-      zone,
-    });
-
-    new route53.AaaaRecord(this, "Alias", {
-      zone,
-      recordName: wwwDomainName,
-      target: route53.RecordTarget.fromAlias(new targets.CloudFrontTarget(cf)),
-    });
-
-    new HttpsRedirect(this, "Redirect", {
-      recordNames: [domainName],
-      targetDomain: wwwDomainName,
-      zone: route53.HostedZone.fromHostedZoneAttributes(this, "HostedZone", {
-        hostedZoneId,
-        zoneName: domainName,
-      }),
-    });
-
-    return fargateService;
+    this.loadbalancer = fargateService.loadBalancer;
   }
 }
